pyproject.toml
[project]
name = "mcp-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "fastmcp>=2.13.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.22.0",
    "python-dotenv>=1.2.1",
]
